Fix warnings
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@12702 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/svc.cpp
@@ -2550,7 +2550,7 @@ static void service_get(Service* service,
 
 			if (temp_len) {
 				/* If data is available, read as much as will fit in buffer */
-				temp_len = MIN(temp_len, length - (buf - buffer));
+				temp_len = MIN(temp_len, (DWORD) (length - (buf - buffer)));
 				const USHORT bytes_read =
 					service_read(service, buf, (USHORT) temp_len, flags);
 				buf += bytes_read;
